requirements.txt
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
